move Model to active_record_spec.rb; fix formatting
